chore(database): setup migrations
