+++ back-end/app.yaml
@@ -19,12 +19,12 @@ env_variables:
   
 
 handlers:
-  # This configures Google App Engine to serve the files in the app's static
+  # This configures Google App Engine to serve the files in the src's static
   # directory.
 - url: /static
   static_dir: static
 
-  # This handler routes all requests not caught above to your main app. It is
+  # This handler routes all requests not caught above to your main src. It is
   # required when static routes are defined, but can be omitted (along with
   # the entire handlers section) when there are no static files defined.
 - url: /.*
