--- back-end/app/index.py
@@ -1,5 +1,0 @@
-from main import app
-
-@app.route('/')
-def hello_world():
-    return 'Hello World!'
