+++ back-end/config.py
@@ -1,0 +1,3 @@
+import os 
+
+PROJECT_ROOT = os.path.dirname(os.path.abspath(__file__))
