+++ back-end/requirements.txt
@@ -2,6 +2,7 @@ aiofiles==25.1.0
 aiohappyeyeballs==2.6.1
 aiohttp==3.13.2
 aiosignal==1.4.0
+alembic==1.17.2
 attrs==25.4.0
 blinker==1.9.0
 cachetools==6.2.2
@@ -13,6 +14,7 @@ cloud-sql-python-connector==1.18.5
 cryptography==46.0.3
 dnspython==2.8.0
 Flask==3.1.2
+Flask-Alembic==3.2.0
 Flask-MySQL==1.6.0
 Flask-SQLAlchemy==3.1.1
 frozenlist==1.8.0
@@ -20,6 +22,7 @@ google-auth==2.43.0
 idna==3.11
 itsdangerous==2.2.0
 Jinja2==3.1.6
+Mako==1.3.10
 MarkupSafe==3.0.3
 multidict==6.7.0
 mysql-connector-python==9.5.0
