+++ back-end/src/models/__init__.py
@@ -1,0 +1,4 @@
+from .test_table import TestTable
+
+
+__all__ = ['TestTable']
